ruteo/requirements.txt
bcrypt==4.0.1
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==40.0.2
et-xmlfile==1.1.0
geographiclib==1.52
geopy==2.2.0
googlemaps==4.10.0
idna==3.4
joblib==1.1.1
numpy==1.19.5
openpyxl==3.1.2
pandas==1.1.5
paramiko==3.3.1
polyline==1.4.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.27.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
threadpoolctl==3.1.0
urllib3==1.26.16
wincertstore==0.2
